pyproject.toml
[project]
name = "langgraph-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "jupyterlab>=4.4.3",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.24",
    "langchain-tavily>=0.2.3",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
